print_listint

+++ 0x13-more_singly_linked_lists/0-print_listint.c
@@ -7,7 +7,7 @@
  */
 size_t print_listint(const listint_t *h)
 {
-	int count = 0;
+	size_t count = 0;
 
 	if (h == NULL)
 		return (0);
